+++ src/pages/Feedback/Feedback.tsx
@@ -16,6 +16,7 @@ import { theme } from '../../styles/theme';
 import Card from './components/Card';
 import LevelModal from './components/LevelModal';
 import FeedbackMemo from '../../shared/components/Feedback/FeedbackMemo';
+import useFollowUpQuestion from './hooks/useFollowUpQuestion';
 
 export interface Question {
   questionId: number;
@@ -61,16 +62,6 @@ export interface Feedback {
   status: string; // "PENDING", "COMPLETED" 등
   content: FeedbackContent;
   updatedAt: string;
-}
-
-interface IFollowUpPayload {
-  message: string;
-  generatedCount: number;
-}
-interface IFollowUpResponse {
-  nextQuestionId: number;
-  questionText: string;
-  // ... 기타 응답 필드
 }
 
 const FeedbackPage = () => {
@@ -132,41 +123,13 @@ const FeedbackPage = () => {
   const { data: feedback } = useFetch<Feedback>(feedbackUrl);
   const { patchData } = usePatch<AnswerPayload, AnswerPayload>(answerUrl);
   console.log('FeedbackPage API 응답 데이터:', data);
+
+  // 꼬리 질문 여부(창목)
   const followUp = data?.followUp;
   console.log('followUp 값:', followUp);
 
-  // 꼬리 질문
-  const [followedQ, setFollowedQ] = useState<IFollowUpResponse | null>(null);
-  const [followedQLoading, setFollowedQLoading] = useState(false);
-  console.log(followedQ);
-  console.log(followedQLoading);
-
-  const [answer, setAnswer] = useState('');
-
-  const handleRequestFollowUp = async () => {
-    const payload: IFollowUpPayload = {
-      message: answer,
-      generatedCount: 1,
-    };
-
-    setFollowedQLoading(true);
-
-    try {
-      const response = await apiClient.post<IFollowUpResponse>(
-        `api/questions/followUp/${id}`,
-        payload
-      );
-
-      setFollowedQ(response.data);
-      console.log('요청 성공:', response.data);
-      setAnswer('');
-    } catch (err) {
-      console.error('요청 실패:', err);
-    } finally {
-      setFollowedQLoading(false);
-    }
-  };
-  // handleRequestFollowUp();
+  // 꼬리 질문 훅 분리(창목)
+  const { followUpQuestion, handleRequestFollowUp } = useFollowUpQuestion(id);
 
   const question = data?.question;
 
@@ -284,10 +247,10 @@ const FeedbackPage = () => {
 
       <FeedbackMemo id={id} memo={data?.memo} />
 
-      {/* TODO: 테스트 해보기 */}
+      {/* TODO: 테스트 해보기(창목) */}
       {!followUp && (
         <QButton onClick={handleRequestFollowUp}>
-          {followedQ === null ? '꼬리 질문 생성' : '꼬리 질문이 생성 되었습니다'}
+          {followUpQuestion === null ? '꼬리 질문 생성' : '꼬리 질문이 생성 되었습니다'}
         </QButton>
       )}
